Performance improvement + interface for using the fact that only one entry changes
If [r,c] is the changed entry and I'm mapping a line, which previous
line lies before r or c, mapped line cannot be mapped after r or c
respectively. The interface prepares the program to remember the whole
building tree from previous run of avoid function thinking that it will
be easier to run avoid function with that knowledge.